# openapi.yml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /upload/:
    post:
      summary: Upload File
      operationId: upload_file_upload__post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_file_upload__post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /files/:
    get:
      summary: Read Files
      operationId: read_files_files__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  '/files/{file_id}':
    delete:
      summary: Delete File
      operationId: delete_file_files__file_id__delete
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: integer
            title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /:
    get:
      summary: Index
      operationId: index__get
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
components:
  schemas:
    Body_upload_file_upload__post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
        - file
      title: Body_upload_file_upload__post
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError